pyproject.toml
[project]
name = "home-hospice-care-ops"
version = "1.0.0"
description = "Home Health & Hospice Operations Dashboard"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "asyncpg>=0.29.0",
    "databricks-sdk>=0.30.0",
    "pydantic>=2.0.0",
    "python-multipart>=0.0.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["server"]
